Rename fancy -> pretty.
