+++ prettyReporter.js
