--- test/jshint.fancy.spec.js
@@ -1,4 +1,0 @@
-require('../index.js')({
-    title: 'fancy reporter',
-    fancy: true
-});
