+++ test/jshint.pretty.spec.js
@@ -1,0 +1,4 @@
+require('../index.js')({
+    title: 'pretty reporter',
+    pretty: true
+});
